re-try delete to compensate possible race conditions (#5843)

+++ redaxo/src/core/lib/util/file.php
@@ -194,9 +194,17 @@ class rex_file
     public static function delete($file)
     {
         return rex_timer::measure(__METHOD__, static function () use ($file) {
-            if (is_file($file)) {
-                return unlink($file);
+
+            $tryUnlink = @unlink($file);
+
+            // re-try without error suppression to compensate possible race conditions
+            if (!$tryUnlink) {
+                clearstatcache(true, $file);
+                if (is_file($file)) {
+                    return unlink($file);
+                }
             }
+
             return true;
         });
     }
